requirements.txt
requests>=2.24.0
Flask>=1.1.2
pycryptodome>=3.9.8